Revert "seperate the shell and api into two files, where shell uses api interface"
This reverts commit e9d5c4c97ad9b15b54450bbf6d5d6f2107bc5e6b.
