+++ Makefile
@@ -2,19 +2,17 @@
 
 # Define variables for paths and compiler
 CC = gcc
-SRC = src/shell.c
-GEN_SRC = src/generation.c
+SRC = src/gst.c
 OUT = bin/intelliShell
 CURL = -lcurl
 JSON = -lcjson
-READLLINE= -lreadline
 
 # Default target to compile and run
 all: $(OUT) run
 
 # Target to compile the program, and link curl
 $(OUT): $(SRC)
-	$(CC) $(SRC) $(GEN_SRC) -o $(OUT) $(CURL) $(JSON) $(READLLINE)
+	$(CC) $(SRC) -o $(OUT) $(CURL) $(JSON)
 
 # Target to run the compiled executable
 run: $(OUT)
