+++ src/shell.c
@@ -7,7 +7,6 @@
 #include <sys/wait.h>
 #include <readline/readline.h>
 #include <readline/history.h>
-#include "generation.h"  // Include the header file
 
 #define MAX_COMMAND_LENGTH 1024
 #define MAX_ARGS 100
@@ -25,19 +24,12 @@ void parse_command(char* command, char** args) {
 }
 
 // Function to execute the command
-void execute_command(char** args, CURL *curl, CURLcode res) {
+void execute_command(char** args) {
     pid_t pid = fork(); // Create a child process
     if (pid == 0) { // Child process
         execvp(args[0], args); // Try to execute the command
         // If execvp returns, an error occurred
-        printf("The following error occurred, generating AI response...\n");
         perror(args[0]); // Print the command name with the error
-        // Loop through all arguments to concat to a string
-        // Call API function to get generation based on failure
-        if (get_generation(args[0], curl, res) != 0) {
-            fprintf(stderr, "Failed to get generation\n");
-            curl_easy_cleanup(curl);
-        }
         exit(EXIT_FAILURE);
     } else if (pid < 0) { // Error forking
         perror("fork");
@@ -52,24 +44,6 @@ int main() {
     char command[MAX_COMMAND_LENGTH];
     char* args[MAX_ARGS];
     char cwd[MAX_PATH];
-    // Initialize CURL library
-    CURL *curl;
-    CURLcode res;
-    curl_global_init(CURL_GLOBAL_DEFAULT);
-    curl = curl_easy_init();
-    
-    // Check if CURL initialization was successful
-    if (!curl) {
-        fprintf(stderr, "Failed to initialize CURL\n");
-        return -1;
-    }
-    
-    // Set the API credential
-    if (set_api_credential() != 0) {
-        fprintf(stderr, "Cannot proceed, as the API key could not be set\n");
-        curl_easy_cleanup(curl);
-        return -1;
-    }
 
     while (1) {
         // Store working directory and print error on failure
@@ -127,12 +101,8 @@ int main() {
         }
 
         // Execute the command
-        execute_command(args, curl, res);
+        execute_command(args);
     }
-
-    // Clean up CURL
-    curl_easy_cleanup(curl);
-    curl_global_cleanup();
 
     return 0;
 }
